scripts: setup-xonsh: fail as soon as a command fails
Setting the -e forces the script to fail as soon as a command fails,
and return the exit status. This is useful to avoid running the
remaining commands when one of them fails and more important to avoid
the front end to show a success message when the script failed.

Signed-off-by: Matheus Castello <matheus.castello@toradex.com>