scripts: Make sure that WSL_DISTRO_NAME is not empty
There is some edge cases that change the flow to use Windows host
tools if a WSL_DISTRO_NAME var is set, but this could be set by the
docker dev container but empty. This patch adds a check to make sure
that the WSL_DISTRO_NAME is not empty before trying to use Windows
host tools.

Signed-off-by: Matheus Castello <matheus.castello@toradex.com>